pyproject.toml
[project]
name = "fifth-symphony"
version = "0.1.0"
description = "AI-powered distributed automation conductor with voice feedback and network orchestration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mlx-whisper>=0.4.3; sys_platform == 'darwin'",
    "numpy>=2.2.6",
    "ollama>=0.6.0",
    "pyaudio>=0.2.14; sys_platform == 'darwin'",
    "textual>=6.2.1",
    "elevenlabs>=1.0.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "rich>=13.0.0",
    "asyncio-throttle>=1.0.2",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "watchdog>=3.0.0",
    # Multi-agent chat dependencies
    "websockets>=12.0",
    # TUI Framework for multi-pane dashboard
    "textual>=0.47.0",
    "textual-dev>=1.2.0",
    # GUI Dependencies
    "PySide6>=6.6.0",
    "pyqtdarktheme>=2.1.0",
    "qasync>=0.24.0",
    # Terminal emulator for embedded console
    "ptyprocess>=0.7.0",
    # Icons and themes
    "qtawesome>=1.3.0",
    # MCP server dependencies
    "mcp>=0.9.0",
    # Docker monitoring
    "docker>=7.0.0",
    # CLI framework
    "click>=8.1.0",
    # Database
    "psycopg[binary]>=3.1.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    # AniList integration
    "AnilistPython>=0.1.3",
    # Speech-to-text
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    # macOS automation (platform-specific)
    "pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-EventKit>=10.0; sys_platform == 'darwin'", # Calendar/Reminders
    "pyobjc-framework-ScriptingBridge>=10.0; sys_platform == 'darwin'", # AppleScript bridge
    # TTS API Server
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "rich-click>=1.9.4",
    # YouTube Subscriptions Integration
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "typer>=0.9.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["modules"]

[project.scripts]
fifth-symphony = "cli:cli"
symphony = "cli:cli"
voice-listener = "modules.voice_listener:main"
dashboard = "modules.dashboard:main"
claude-monitor = "modules.claude_monitor:main"
agent-monitor = "modules.agent_monitor:main"
youtube-subs = "modules.youtube_subscriptions:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pylint>=3.3.8",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.12.10",
    "defusedxml>=0.7.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".venv",
    "venv",
    "old-scripts",
    "__pycache__",
    ".git",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C", "Q"]  # Core quality rules
ignore = [
    "E402",    # module-import-not-at-top-of-file (import after dependency check)
    "E501",    # line-too-long (handled by formatter)
    "E722",    # bare-except (intentional for resilient error handling)
    "N802",    # invalid-function-name (Qt methods)
    "N812",    # lowercase-imported-as-non-lowercase
    "BLE001",  # blind-except (sometimes necessary)
    "C901",    # complex-structure (sometimes unavoidable)
    "G004",    # logging-f-string (f-strings are more readable)
    "S603",    # subprocess-without-shell-equals-true (secure usage)
    "S607",    # start-process-with-partial-path (scripts are trusted)
    "S311",    # suspicious-non-cryptographic-random-usage (not for crypto)
    "S110",    # try-except-pass (sometimes intentional)
    "ARG002",  # unused-method-argument (GUI callbacks)
    "FBT001",  # boolean-type-hint-positional-argument
    "FBT002",  # boolean-default-value-positional-argument
    "FBT003",  # boolean-positional-value-in-call
    "DTZ005",  # call-datetime-now-without-tzinfo (local time is fine)
    "DTZ006",  # call-datetime-fromtimestamp (local time is fine)
    "ASYNC221", # run-process-in-async-function (intentional)
    "ASYNC230", # blocking-open-call-in-async-function (quick reads)
    "B007",    # unused-loop-control-variable (enumerate with unused index)
    "B019",    # cached-instance-method (lru_cache on methods is fine)
    "PLC0415", # import-outside-top-level (conditional imports)
    "EXE001",  # shebang-not-executable (scripts might be run via interpreter)
]

[tool.black]
line-length = 100
target-version = ["py310"]
